Provide a convenient `replication-delay` method in client
Replication delay is the de-facto metric used by freno to determine
whether read or write access to a db cluster should be throttled.

Despite the following is said in the
[docs](https://github.com/github/freno/blob/master/doc/http.md#get-method):

> GET and HEAD respond with same status codes. But GET requests compute
and return additional data. Automated requests should not be interested
in this data; the status code is what should guide the clients. However
humans or manual requests may benefit from extra information supplied by
the GET request.

A computer might also be interested in knowing the raw value of
replication delay to, for instance, route DB operations to replicas
depending on when the last write happened on a primary server.