Fix regression from capturing StopIteration
We encountered a bug when upgradaing where loop captures the
StopIteration exception, whereas a plain while loop does not. This
commit fixes this regression by breaking from the loop before yielding.

It's better for us to break out of the loop anyways before yielding both
to make control flow more obvious and to remove a confusing loop frame
from stacktraces.

Co-authored-by: Matthew Draper <matthew@trebex.net>