Revert Rubocop changes to throttle loop
Following Style/InfiniteLoop introduced a regression because it's an
unsafe change, so we should not apply this in the future. There's no
problem with while loops and we should be comfortable using them.
Style/Next didn't cause a behaviour change, but the resulting code was
much harder to read.

Rubocop's rules aren't inherently virtuous, IMO we should write code in
a way that works the best and is the best for humans to read rather than
following an arbitrary set of rules. Given that these two rules failed
us here, let's not attempt to apply them elsewhere either.